# openapi.yaml
openapi: 3.0.3
info:
  title: Sports Companion API
  version: 1.0.0
servers:
  - url: https://api.sportification.app/v1
paths:
...
  /venues:
    get:
      summary: List venues
      responses:
        '200':
          description: Successful response
    post:
      summary: Create venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venue'
      responses:
        '201':
          description: Venue created
  /venues/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      summary: Get venue
      responses:
        '200':
          description: Venue data
    patch:
      summary: Update venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Venue'
      responses:
        '200':
          description: Venue updated
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
...
    Venue:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        location: { type: object }
        surfaceType: { type: string }
        capacity: { type: integer }
...